Use FileDescriptor and frysk.rsl in frysk.testbed.
frysk-core/frysk/ftrace/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestLtrace.java: Update to match frysk.testbed.
	* TestMappingGuard.java: Ditto.

frysk-core/frysk/proc/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestTaskTerminateObserver.java: Update to match frysk.testbed.
	* TestTaskSyscallObserver.java: Update.
	* TestProcGet.java: Ditto.
	* TestExec.java: Ditto.
	* TestTaskClonedObserver.java: Ditto.
	* TestTaskObserverDetach.java: Ditto.
	* TestFindProc.java: Ditto.

frysk-core/frysk/proc/live/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestRefresh.java: Update to match frysk.testbed.
	* TestTaskObserverCode.java: Ditto.
	* TestRuntimeIsa.java: Ditto.

frysk-core/frysk/testbed/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	Replace "int pid" with ProcessIdentifier; use frysk.rsl.
	* TestLib.java: Update.
	* TestfileTokenScanner.java: Update.
	* TaskObserverBase.java: Update.
	* SignalWaiter.java: Update.
	* Offspring.java (getPid()): Return a ProcessIdentifier.
	(logger): Delete.
	* ExecOffspring.java: Update.
	* SynchronizedOffspring.java: Update.
	* SlaveOffspring.java: Update.
	(assertSendAddForkWaitForAcks(int)): Delete.
	(assertSendAddCloneWaitForAcks(int)): Delete.
	* LegacyOffspring.java: Update.
	(assertSendAddForkWaitForAcks(int)): Delete.
	(assertSendAddCloneWaitForAcks(int)): Delete.
	* StopEventLoopWhenProcRemoved.java: Update.
	* StatState.java: Update.
	(assertRunUntil(Task)): New.
	(assertIs(Task)): New.
	* DaemonBlockedAtSignal.java: Extend Offspring; update.
	* DaemonBlockedAtEntry.java: Ditto; update.

frysk-sys/frysk/rsl/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* Log.java: Add more loggers.

+++ frysk-core/frysk/proc/TestFindProc.java
@@ -1,6 +1,6 @@
 // This file is part of the program FRYSK.
 //
-// Copyright 2005, Red Hat Inc.
+// Copyright 2005, 2007, 2008, Red Hat Inc.
 //
 // FRYSK is free software; you can redistribute it and/or modify it
 // under the terms of the GNU General Public License as published by
@@ -120,8 +120,8 @@ public class TestFindProc
      * Find out how many processes are associated with the test process.
      * Should be just the one.
      */
-    FindProc finder = new MyFinder(ackProc.getPid());
-    Manager.host.requestProc(ackProc.getPid(), finder);
+    FindProc finder = new MyFinder(ackProc.getPid().intValue());
+    Manager.host.requestProc(ackProc.getPid().intValue(), finder);
     assertRunUntilStop("testFindProc");
 
     int postFind = o.getCount();
